third_party/blink/web_tests/platform/generic/fast/css/fontfaceset-set-operations-expected.txt

Tests Set operations of FontFaceSet.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.fonts.size is 3
check forEach
PASS callbackArgs[0] is an instance of FontFace
PASS callbackArgs[0] === callbackArgs[1] is true
PASS callbackArgs[2] === document.fonts is true
PASS thisValue === thisArg is true
PASS faces.length is 3
PASS faces[0].family is "Font1"
PASS faces[1].family is "Font2"
PASS faces[2].family is "Font3"
PASS document.fonts.has(faces[0]) is true
PASS document.fonts.has(faces[1]) is true
PASS document.fonts.has(faces[2]) is true
check keys
PASS faces.length is 3
PASS faces[0].family is "Font1"
PASS faces[1].family is "Font2"
PASS faces[2].family is "Font3"
PASS document.fonts.has(faces[0]) is true
PASS document.fonts.has(faces[1]) is true
PASS document.fonts.has(faces[2]) is true
check values
PASS faces.length is 3
PASS faces[0].family is "Font1"
PASS faces[1].family is "Font2"
PASS faces[2].family is "Font3"
PASS document.fonts.has(faces[0]) is true
PASS document.fonts.has(faces[1]) is true
PASS document.fonts.has(faces[2]) is true
check entries
PASS entry[0] === entry[1] is true
PASS entry[0] === entry[1] is true
PASS entry[0] === entry[1] is true
PASS faces.length is 3
PASS faces[0].family is "Font1"
PASS faces[1].family is "Font2"
PASS faces[2].family is "Font3"
PASS document.fonts.has(faces[0]) is true
PASS document.fonts.has(faces[1]) is true
PASS document.fonts.has(faces[2]) is true
PASS document.fonts.has(null) threw exception TypeError: Failed to execute 'has' on 'FontFaceSet': parameter 1 is not of type 'FontFace'..
PASS document.fonts.has("Font1") threw exception TypeError: Failed to execute 'has' on 'FontFaceSet': parameter 1 is not of type 'FontFace'..
PASS document.fonts.add(null) threw exception TypeError: Failed to execute 'add' on 'FontFaceSet': parameter 1 is not of type 'FontFace'..
PASS document.fonts.add("Font1") threw exception TypeError: Failed to execute 'add' on 'FontFaceSet': parameter 1 is not of type 'FontFace'..
PASS document.fonts.delete(null) threw exception TypeError: Failed to execute 'delete' on 'FontFaceSet': parameter 1 is not of type 'FontFace'..
PASS document.fonts.delete("Font1") threw exception TypeError: Failed to execute 'delete' on 'FontFaceSet': parameter 1 is not of type 'FontFace'..
PASS document.fonts.has(nonCssConnectedFace) is false
PASS document.fonts.add(faces[0]) did not throw exception.
PASS document.fonts.size is 3
PASS document.fonts.delete(faces[0]) is false
PASS document.fonts.size is 3
PASS document.fonts.size is 4
PASS document.fonts.has(nonCssConnectedFace) is true
PASS document.fonts.size is 4
PASS document.fonts.size is 3
PASS document.fonts.has(nonCssConnectedFace) is false
PASS document.fonts.size is 3
PASS document.fonts.size is 4
PASS document.fonts.size is 3
PASS successfullyParsed is true

TEST COMPLETE
